Move GGUF metadata scanning into model-artifact
